+++ view/forum/topicPosts.php
@@ -20,11 +20,17 @@ if($posts ){
  foreach($posts as $post){ ?>
     <div class="flex" >
 
-        <p><?=  htmlspecialchars_decode($post->getText())  ?></p> 
-        <a style="color:red ;" href="index.php?ctrl=forum&action=deletePost&id=<?= $post->getId() ?>">Supprimer ce post</a>
-        <button onclick=" if( document.querySelector('.FormTopic<?=$x?>').classList.contains('FormActive') ){
-        document.querySelector('.FormTopic<?=$x?>').classList.remove('FormActive')} else {
-        document.querySelector('.FormTopic<?=$x?>').classList.add('FormActive') }">modifier</button>
+        <p><?=  htmlspecialchars_decode($post->getText())  ?></p>  
+        
+        <?php if ($post->MadeBy($_SESSION["user"]) ) {?>
+
+            <a style="color:red ;" href="index.php?ctrl=forum&action=deletePost&id=<?= $post->getId() ?>">Supprimer ce post</a>
+
+            <button onclick=" if( document.querySelector('.FormTopic<?=$x?>').classList.contains('FormActive') ){
+            document.querySelector('.FormTopic<?=$x?>').classList.remove('FormActive')} else {
+            document.querySelector('.FormTopic<?=$x?>').classList.add('FormActive') }">modifier</button> 
+
+        <?php } ?>
 
     </div>
 
@@ -39,12 +45,20 @@ if($posts ){
 <?php  $x+=1; }
 }
 else { echo " il n'y aucun post dans ce topic ";} 
-if ($topic->getFermer()==0){ ?>
+if ($topic->getFermer()==0){ 
+    if(App\Session::getUser()){?>
+        <form method="post" action="index.php?ctrl=forum&action=nvPost&id=<?= $topic->getId() ?>">
+        <p><label>Text</label>
+            <input type="text" name="text" required></p>
 
-<form method="post" action="index.php?ctrl=forum&action=nvPost&id=<?= $topic->getId() ?>">
-<p><label>Text</label>
-    <input type="text" name="text" required></p>
+            <input  type="submit" name="submit" value="submit">
+        </form>
+    <?php } else { ?> 
+    <p>
+        <a href="index.php?ctrl=security&action=login">Connecter</a>
+        ou 
+        <a href="index.php?ctrl=security&action=register"> Inscriver</a>
+        vous pour ajouter un post !
 
-    <input  type="submit" name="submit" value="submit">
-</form>
-<?php } else { ?> <p>Ce Topic est Verrouiller impossible d'ajouter des posts </p><?php }
+    </p> <?php }
+} else { ?> <p>Ce Topic est Verrouiller impossible d'ajouter des posts </p><?php }
